Fix: UI and functionality bugs
- Restored the previous card layout.
- Fixed the main navigation menu to be at the top of the page.
- Fixed the card click functionality by loading the scripts in the correct order.
- Fixed the language and theme toggles by moving the event listeners to `fabmain.js`.
