+++ global.css
@@ -36,6 +36,10 @@ body.dark {
   justify-content: space-between;
   padding: 1.2rem 2rem;
   background: transparent;
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
 }
 .ops-nav .ops-logo {
   font-weight: bold;
@@ -88,8 +92,8 @@ body.dark {
   margin-inline: auto;
   margin-top: 2.8rem;
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(min(19rem, 100%), 1fr));
-  gap: 2.4rem;
+  grid-template-columns: repeat(auto-fit, minmax(min(15rem, 100%), 1fr));
+  gap: 1.5rem;
 }
 .card {
   padding: 2.3rem;
@@ -276,8 +280,8 @@ footer {
   background: var(--clr-footer);
   color: #fff;
   font-size: 1rem;
-  text-align: right;
-  padding: 1.1rem 2.5rem 1.1rem 0.8rem;
+  text-align: center;
+  padding: 1.1rem 0;
   position: fixed;
   left: 0;
   bottom: 0;
